requirements.txt
Flask==3.1.2
libvirt-python==11.10.0
